refactor: Moved ansi escape length functions to own package
